chore: move unit test in common folder
